# src/routes/food.js
const express = require('express');
const router = express.Router();
const { listFoodItems, createFoodItem, getFoodItem, updateFoodItem, deleteFoodItem } = require('../models/store');

router.get('/', async (req, res) => {
  const list = await listFoodItems();
  res.json(list);
});

router.get('/:id', async (req, res) => {
  const f = await getFoodItem(req.params.id);
  if (!f) return res.status(404).json({ error: 'Food item not found' });
  res.json(f);
});

router.post('/', async (req, res) => {
  const { name, price, description, category } = req.body;
  if (!name || price == null) return res.status(400).json({ error: 'name and price are required' });
  const created = await createFoodItem({ name, price, description, category });
  res.status(201).json(created);
});

router.put('/:id', async (req, res) => {
  const updated = await updateFoodItem(req.params.id, req.body);
  if (!updated) return res.status(404).json({ error: 'Food item not found' });
  res.json(updated);
});

router.delete('/:id', async (req, res) => {
  const ok = await deleteFoodItem(req.params.id);
  if (!ok) return res.status(404).json({ error: 'Food item not found' });
  res.status(204).send();
});

module.exports = router;